+++ Data/Star.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP               #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Data.Star
@@ -19,7 +20,9 @@ import Data.Semiring
 --
 -- @'aplus' x = x '*' 'star' x@
 class (Semiring a) => Star a where
-  {-# MINIMAL star | aplus #-} 
+#if __GLASGOW_HASKELL__ >= 708  
+  {-# MINIMAL star | aplus #-}
+#endif
   star :: a -> a
   star a = one `plus` aplus a